Module Buffered9PReader.Make

Parameters

module Log : sig ... end

module FLOW : Mirage_flow.S

Signature

type t

A buffered 9P message reader over a FLOW

val create : FLOW.flow -> t

Construct a buffered 9P message reader over a FLOW

val read : t -> (Cstruct.t, [ `Msg of string ]) Stdlib.result Lwt.t

Return the unmarshalled body of the next message read from the flow, without the initial length field.