Module type Filesystem.S

A filesystem implementation

type t

A filesystem. This is likely shared amongst many connections.

type connection

Server state associated with a particular connection. This should contain connection-local data such as cached authentication information and a private space of Fids

val connect : t -> Protocol_9p__.Protocol_9p_info.t -> connection

Called after making a connection to initialise the per-connection state

val attach : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Attach.t -> Protocol_9p__.Protocol_9p_response.Attach.t or_error Lwt.t

val walk : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Walk.t -> Protocol_9p__.Protocol_9p_response.Walk.t or_error Lwt.t

val clunk : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Clunk.t -> unit or_error Lwt.t

val open_ : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Open.t -> Protocol_9p__.Protocol_9p_response.Open.t or_error Lwt.t

val read : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Read.t -> Protocol_9p__.Protocol_9p_response.Read.t or_error Lwt.t

val stat : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Stat.t -> Protocol_9p__.Protocol_9p_response.Stat.t or_error Lwt.t

val create : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Create.t -> Protocol_9p__.Protocol_9p_response.Create.t or_error Lwt.t

val write : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Write.t -> Protocol_9p__.Protocol_9p_response.Write.t or_error Lwt.t

val remove : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Remove.t -> unit or_error Lwt.t

val wstat : connection -> cancel:unit Lwt.t -> Protocol_9p__.Protocol_9p_request.Wstat.t -> unit or_error Lwt.t