Module Protocol_9p.Types

Parsers and printers for types used in 9P messages

val big_enough_for : string -> Cstruct.t -> int -> (unit, [ `Msg of string ]) Stdlib.result

big_enough_for name buf length returns an error with a log message if buffer buf is smaller than length. The name will be included in the error message.

module Int8 : sig ... end

module Int16 : sig ... end

module Int32 : sig ... end

module Int64 : sig ... end

module Version : sig ... end

module Fid : sig ... end

module OpenMode : sig ... end

module FileMode : sig ... end

module Qid : sig ... end

module Tag : sig ... end

module Data : sig ... end

module Stat : sig ... end

module Arr (T : sig ... end) : sig ... end

A sequence of T.ts written contiguously to a buffer