Module Types.Int32

include module type of Stdlib.Int32

val zero : int32

val one : int32

val minus_one : int32

val neg : int32 -> int32

val add : int32 -> int32 -> int32

val sub : int32 -> int32 -> int32

val mul : int32 -> int32 -> int32

val div : int32 -> int32 -> int32

val unsigned_div : int32 -> int32 -> int32

val rem : int32 -> int32 -> int32

val unsigned_rem : int32 -> int32 -> int32

val succ : int32 -> int32

val pred : int32 -> int32

val abs : int32 -> int32

val max_int : int32

val min_int : int32

val logand : int32 -> int32 -> int32

val logor : int32 -> int32 -> int32

val logxor : int32 -> int32 -> int32

val lognot : int32 -> int32

val shift_left : int32 -> int -> int32

val shift_right : int32 -> int -> int32

val shift_right_logical : int32 -> int -> int32

val of_int : int -> int32

val to_int : int32 -> int

val unsigned_to_int : int32 -> int option

val of_float : float -> int32

val to_float : int32 -> float

val of_string : string -> int32

val of_string_opt : string -> int32 option

val to_string : int32 -> string

val bits_of_float : float -> int32

val float_of_bits : int32 -> float

type t = int32

val compare : t -> t -> int

val unsigned_compare : t -> t -> int

val equal : t -> t -> bool

val min : t -> t -> t

val max : t -> t -> t

val format : string -> int32 -> string

deprecated Use Printf.sprintf with a [%l...] format instead.

val t_of_sexp : Sexplib.Sexp.t -> t

val sexp_of_t : t -> Sexplib.Sexp.t

val any : t

val is_any : t -> bool

val sizeof : t -> int

The size of a buffer needed to hold t

val read : Cstruct.t -> (t * Cstruct.t, [ `Msg of string ]) Stdlib.result

Read a t from the given buffer and return it, along with the unused remainder of the buffer. If the buffer cannot be parsed then return an error.

val write : t -> Cstruct.t -> (Cstruct.t, [ `Msg of string ]) Stdlib.result

Write a t into the given buffer. If the buffer is too small, then return an error. Return the unused remainder of the buffer.